requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==24.2.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
datasets==4.4.2
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
Django==4.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.59.0
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.7.0
future==1.0.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.2
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lark==1.2.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
openai==1.97.1
opencv-python==4.12.0.88
orjson==3.11.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymodbus==3.7.4
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.13
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
seaborn==0.13.2
selenium==4.12.0
Send2Trash==1.8.3
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.43
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.47.2
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchcodec==0.9.1
torchvision==0.23.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
